Consensus: MOVEONLY: Move CheckBlockHeader, IsSuperMajority and ContextualCheckBlockHeader() to consensus.o
