+++ src/consensus/consensus.cpp
@@ -1,0 +1,52 @@
+// Copyright (c) 2009-2010 Satoshi Nakamoto
+// Copyright (c) 2009-2014 The Bitcoin Core developers
+// Distributed under the MIT software license, see the accompanying
+// file COPYING or http://www.opensource.org/licenses/mit-license.php.
+
+#include "consensus/consensus.h"
+
+#include "chain.h"
+#include "consensus/validation.h"
+
+bool Consensus::CheckBlockHeader(const CBlockHeader& block, int64_t nTime, CValidationState& state, const Consensus::Params& params, bool fCheckPOW)
+{
+    // Check proof of work matches claimed amount
+    if (fCheckPOW && !CheckProofOfWork(block.GetHash(), block.nBits, params))
+        return state.DoS(50, false, REJECT_INVALID, "high-hash");
+
+    // Check timestamp
+    if (block.GetBlockTime() > nTime + 2 * 60 * 60)
+        return state.Invalid(false, REJECT_INVALID, "time-too-new");
+
+    return true;
+}
+
+bool IsSuperMajority(int minVersion, const CBlockIndex* pstart, unsigned nRequired, unsigned nToCheck)
+{
+    unsigned int nFound = 0;
+    for (unsigned int i = 0; i < nToCheck && nFound < nRequired && pstart != NULL; i++)
+    {
+        if (pstart->nVersion >= minVersion)
+            ++nFound;
+        pstart = pstart->pprev;
+    }
+    return (nFound >= nRequired);
+}
+
+bool Consensus::ContextualCheckBlockHeader(const CBlockHeader& block, CValidationState& state, CBlockIndex* pindexPrev, const Consensus::Params& params)
+{
+    // Check proof of work
+    if (block.nBits != GetNextWorkRequired(pindexPrev, &block, params))
+        return state.DoS(100, false, REJECT_INVALID, "bad-diffbits");
+
+    // Check timestamp against prev
+    if (block.GetBlockTime() <= pindexPrev->GetMedianTimePast())
+        return state.Invalid(false, REJECT_INVALID, "time-too-old");
+
+    // Reject block.nVersion=n blocks when 95% (75% on testnet) of the network has upgraded, last version=3:
+    for (unsigned int i = 2; i <= 3; i++)
+        if (block.nVersion < 2 && IsSuperMajority(2, pindexPrev, params.nMajorityRejectBlockOutdated, params.nMajorityWindow))
+            return state.Invalid(false, REJECT_OBSOLETE, strprintf("bad-version nVersion=%d", i-1));
+
+    return true;
+}
