+++ src/consensus/consensus.h
@@ -36,4 +36,10 @@ bool ContextualCheckBlockHeader(const CBlockHeader& block, CValidationState& sta
 
 } // namespace Consensus
 
+/**
+ * Returns true if there are nRequired or more blocks of minVersion or above
+ * in the last nToCheck blocks, starting at pstart and going backwards.
+ */
+bool IsSuperMajority(int minVersion, const CBlockIndex* pstart, unsigned nRequired, unsigned nToCheck);
+
 #endif // BITCOIN_CONSENSUS_CONSENSUS_H
